Fix numpy and Cython typing for numpy >= 1.25

+++ CTL/pehe_tree.py
@@ -42,9 +42,9 @@ class PEHETree(_CausalTree):
 
     def fit(self, x, y, t):
         self.column_num = x.shape[1]
-        x = x.astype(np.float)
-        y = y.astype(np.float)
-        t = t.astype(np.float)
+        x = x.astype(np.float64)
+        y = y.astype(np.float64)
+        t = t.astype(np.float64)
         self.tree.fit(x, y, t)
         self.fitted = True
         self.tree_depth = self.tree.tree_depth
